+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Unreleased
 
 ## Added
+- `kaocha.runner/exec` for use with Clojure CLI's -X feature
 
 ## Fixed
 
 - Fix only considering public vars when building up the test plan
+=======
 
 # 1.0.887 (2021-09-01 / 38470aa)